move ChatItem to tailwind
