+++ chat-react2/src/components/Home/Main/ChatHeader/ChatHeader.scss
@@ -1,3 +1,20 @@
+.user__no-img {
+  display: flex;
+  text-transform: uppercase;
+  width: 60px;
+  height: 60px;
+  border-radius: 50%;
+  justify-content: center;
+  background-color: #87ceeb;
+
+  align-items: center;
+
+  & .name {
+    font-weight: 700;
+    color: white;
+  }
+}
+
 .last__activity {
   unicode-bidi: plaintext;
   overflow: hidden;
